Break AcceptedUserTable & InvitedUserTable into their own files
